Adding blur to _events
this is to register the blur for the keyboard.js keyIsDown function #411 fucntion. it will empty the array that contains all the keys currently down when the user no longer has focus on the window. We need this because the keyup event will not be fired when the focus has been changed.